Move Pow to pow module.
