+++ src/ops/mod.rs
@@ -2,4 +2,3 @@ pub mod saturating;
 pub mod checked;
 pub mod wrapping;
 pub mod inv;
-pub mod pow;
